requirements/base.txt
-i https://pypi.org/simple

django==3.0.7
djangorestframework==3.11.0
psycopg2-binary==2.8.6
django-filter==2.3.0
